LPS-56224 - Fix of imports to class "BaseDDMFormDeserializerTestCase" due the renaming of package "/test/unit/dynamicdatamapping".
