+++ modules/apps/dynamic-data-mapping/dynamic-data-mapping-test/test/unit/com/liferay/portlet/dynamic/data/mapping/io/DDMFormJSONDeserializerTest.java
@@ -16,7 +16,7 @@ package com.liferay.portlet.dynamic.data.mapping.io;
 
 import com.liferay.portal.kernel.exception.PortalException;
 import com.liferay.portal.kernel.util.LocaleUtil;
-import com.liferay.portlet.dynamicdatamapping.io.BaseDDMFormDeserializerTestCase;
+import com.liferay.portlet.dynamic.data.mapping.io.BaseDDMFormDeserializerTestCase;
 import com.liferay.portlet.dynamicdatamapping.io.DDMFormJSONDeserializerImpl;
 import com.liferay.portlet.dynamicdatamapping.io.DDMFormJSONDeserializerUtil;
 import com.liferay.portlet.dynamicdatamapping.model.DDMForm;
